update : changes in loginNavigate and file name
