+++ src/atoms/Link/Link.js
@@ -2,9 +2,7 @@ import React from 'react'
 import './link.css'
 const Link = (props) => {
   return (
-    <div className='link'>
-      <a href='#/'>{props.text}</a>
-    </div>
+    <a href={props.link || '#'} className={props.classAddOn || 'link'} >{props.text}</a>
   )
 }
