+++ src/atoms/Link/link.css
@@ -1,7 +1,27 @@
-.link a {
-  color: white;
+@import '../../../src/assets/css/variable.css';
+.link {
+  color: var(--primary-color);
   font-size: 10px;
   font-weight: 400;
   margin-bottom: 50px;
   text-decoration: none;
+}
+.log-in {
+  color: var(--primary-color);
+  font-size: 12px;
+  padding-left: 3px;
+  font-weight: 600;
+  text-decoration: none;
+  @media (max-width: 1800px) {
+    font-size: 20px;
+  }
+  @media (max-width: 1500px) {
+    font-size: 18px;
+  }
+  @media (max-width: 1200px) {
+    font-size: 16px;
+  }
+  @media (max-width: 900px) {
+    font-size: 14px;
+  }
 }
