+++ src/atoms/Text/Text.js
@@ -1,13 +1,8 @@
 import React from 'react'
 import './text.css'
 const Text = (props) => {
-  const isHeading = props.heading
   return (
-    <div className='text'>
-      <div className={isHeading === true ? 'form-heading' : 'normal-text'}>
-        <span>{props.text}</span>
-      </div>
-    </div>
+    <span className={props.classAddOn || 'normal-text'}>{props.text}</span>
   )
 }
