+++ src/atoms/backgroundWithLogo/BackGround.js
@@ -1,12 +1,9 @@
 import React from 'react'
-import profile from '../../../src/assets/images/Profile.svg'
 import './background.css'
 
-const BackGround = (props) => {
-  let isLogo = props.isLogo
+const BackGround = () => {
   return (
     <div className='login-form'>
-      {isLogo && <img src={profile} alt='profile-img' className='profile-img' />}
     </div>
   )
 }
