+++ src/atoms/navigator/navigator.css
@@ -7,7 +7,6 @@
   );
   border-radius: 35px 35px 0px 0px;
   background-size: cover;
-  width: auto;
   padding: 20px 20px 30px 20px;
   @media (max-width: 1800px) {
     margin-top: 150px;
@@ -25,13 +24,15 @@
     margin-top: 120px;
   }
   @media (max-width: 600px) {
-    margin-top: 90px;
+    margin-top: 20px;
+  }
+  @media (max-width: 400px) {
+    margin-top: 10px;
   }
 }
-.already-subscribe-login a {
+
+.already-subscribe {
   color: var(--primary-color);
-}
-.already-subscriber {
   font-weight: 200;
   text-decoration: none;
   font-size: 10px;
@@ -46,23 +47,5 @@
   }
   @media (max-width: 900px) {
     font-size: 12px;
-  }
-}
-
-.already-subscribe-login a span {
-  font-size: 12px;
-  padding-left: 3px;
-  font-weight: 600;
-  @media (max-width: 1800px) {
-    font-size: 20px;
-  }
-  @media (max-width: 1500px) {
-    font-size: 18px;
-  }
-  @media (max-width: 1200px) {
-    font-size: 16px;
-  }
-  @media (max-width: 900px) {
-    font-size: 14px;
   }
 }
