+++ src/atoms/seperator/Seperator.js
@@ -4,7 +4,6 @@ import './seperator.css'
 const Seperator = (props) => {
   return (
     <div className='seperator'>
-
       <p className='or-text'>{props.text}</p>
     </div>
   )
