+++ src/atoms/seperator/seperator.css
@@ -9,10 +9,10 @@
     margin: 60px 0px;
   }
   @media (max-width: 600px) {
-    margin: 50px 0px;
+    margin: 30px 0px;
   }
   @media (max-width: 400px) {
-    margin: 40px 0px;
+    margin: 20px 0px;
   }
 }
 p {
