--- src/atoms/shadow/Shadow.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import './shadow.css'
-const Shadow = () => {
-  return (
-    <div className='background-shadow-img'>
-    </div>
-  )
-}
-
-export default Shadow
