+++ src/molecules/LoginSocialMedia/LoginSocialMedia.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import google from '../../../src/assets/images/Google.svg'
+import facebook from '../../../src/assets/images/Facebook.svg'
+import Button from '../../atoms/button/Button'
+import './loginsocialmedia.css'
+
+const LoginResourse = () => {
+  return (
+    <div className='login-outside-resource'>
+      <Button img={google} text='Continue with Google' alt='google-img' classAddOn1='login-buttons' classAddOn2='login-form-btn' />
+      <Button img={facebook} text='Continue with Facebook' alt='facebook-img' classAddOn1='login-buttons' classAddOn2='login-form-btn' />
+    </div>
+  )
+}
+
+export default LoginResourse
