+++ src/molecules/form/Form.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import Input from '../../atoms/inputs/Input'
-import './logincustominput.css'
-import Button from '../../atoms/buttons/Button'
+import Input from '../../atoms/input/Input'
+import './form.css'
+import Button from '../../atoms/button/Button'
 import {Formik, Form} from 'formik';
 import * as Yup from 'yup';
 
 
-const LoginCustomInput = () => {
+const FormFields = () => {
   const handleSubmit = (values) => {
     console.log(values);
   };
@@ -37,4 +37,4 @@ const LoginCustomInput = () => {
   )
 }
 
-export default LoginCustomInput
+export default FormFields
