+++ src/molecules/policyAndLogin/PolicyAndLogin.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import LoginNavigate from '../../atoms/alreadySubscribe/LoginNavigate'
+import Navigator from '../../atoms/navigator/Navigator'
 import Text from '../../atoms/Text/Text'
 import Link from '../../atoms/Link/Link'
 import './policy-login.css'
@@ -9,9 +9,9 @@ const PolicyAndLogin = () => {
     <>
       <div className='terms-agree-documents'>
         <Text text='By using the app, I agree to The Sport' heading={false} />
-        <Link text='Terms of Services & Privacy Policy' />
+        <Link text='Terms of Services & Privacy Policy' link="#" />
       </div>
-      <LoginNavigate text='Already a Subscriber?' linkText='Log In' />
+      <Navigator text='Already a Subscriber?' linkText='Log In' link='#' />
     </>
   )
 }
