+++ src/organics/loginForm/loginform.css
@@ -12,6 +12,13 @@
     padding-top: 250px;
   }
   @media (max-width: 400px) {
-    padding-top: 230px;
+    padding-top: 210px;
   }
+}
+.profile-img {
+  position: absolute;
+  top: 18%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 1;
 }
