+++ src/template/loginTemplate/LoginTemplate.js
@@ -1,19 +1,16 @@
 import React from 'react'
-import LoginPageOrganic from '../../organics/loginForm/LoginPageOrganic'
+import LoginForm from '../../organics/loginForm/LoginForm'
 import './logintemplate.css'
-import Shadow from '../../atoms/shadow/Shadow'
 
 const LoginTemplate = () => {
   return (
     <div className='screen'>
       <div className='login'>
-        {/* <div className='background-shadow-img'>
-        </div> */}
-        <Shadow />
-        <LoginPageOrganic />
+        <div className='background-shadow-img'>
+        </div>
+        <LoginForm />
       </div>
     </div>
   )
 }
-
 export default LoginTemplate
